pyproject.toml
[project]
name = "bilibili-gachi-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apscheduler>=3.11.1",
    "bilibili-api-python>=17.4.1",
    "bs4>=0.0.2",
    "curl-cffi>=0.15.0",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "pydantic>=2.13.0",
    "setuptools>=82.0.1",
    "tenacity>=9.1.2",
]
